rename mmp::periph*->scc::tlm_target_bfs (bitfield support); adapt to scc licensing; cleanup
