--- incl/mmp/peripheral_base.h
@@ -1,142 +1,0 @@
-/**
- * @Author: sharif
- * @Date:   2019-03-10T02:03:41+01:00
- * @Email:  uzair.sharif@tum.de
- * @Filename: peripheral_base.h
- * @Last modified by:   Johannes Geier (johannes.geier@tum.de)
- * @Last modified time: 2021-01-21T18:11:00+02:00
- * @Copyright: Copyright (c) 2019
-  Chair of Electronic Design Automation, TU Munich
-
-  The above copyright notice and this permission notice shall be included in
-  all copies or substantial portions of the Software.
-
-  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-  THE SOFTWARE.
-
- */
-
-#ifndef __MMP_PERIPHERAL_BASE_H__
-#define __MMP_PERIPHERAL_BASE_H__
-
-#include <memory>
-#include <utility>
-#include <vector>
-
-#include "systemc"
-#include "tlm_utils/simple_target_socket.h"
-#include "scc/tlm_target.h"
-
-#include "util/ities.h"
-
-#define ID_MMP_PERIPHERAL "mmp: Peripheral"
-
-namespace mmp {
-
-struct PerParams {
-  uint64_t base_addr{0};
-  size_t size{0};
-  size_t num_irqs{0};
-  size_t num_regs{0};
-  PerParams( uint64_t base_addr, size_t size, size_t num_irqs, size_t num_regs)
-  : base_addr(base_addr)
-  , size(size)
-  , num_irqs(num_irqs)
-  , num_regs(num_regs) {}
-};
-
-template <class owner_t>
-class PeripheralBase {
- public:
-  sc_core::sc_in<bool> rst_in_{"reset_in"};
-  std::unique_ptr<std::vector<sc_core::sc_out<bool>>> irq_out_{nullptr};
-
-  PeripheralBase(PerParams&& params, owner_t* owner = nullptr)
-    : params_{std::move(params)}
-    , owner_{owner} {
-    irq_out_ = util::make_unique<std::vector<sc_core::sc_out<bool>>>(params_.num_irqs);
-  }
-  virtual ~PeripheralBase() = default;
-
-  const owner_t* getOwner() const { return owner_; }
-
-  void bindIRQ(size_t num, sc_core::sc_signal<bool>* sig) {
-    if (num >= irq_out_->size()) {
-      SC_REPORT_FATAL(ID_MMP_PERIPHERAL, "not enough IRQs in Per::connectIRQ()");
-    }
-
-    (*irq_out_)[num].bind(*sig);
-  }
-
- protected:
-  const PerParams params_{};
-  owner_t* const owner_{nullptr};
-};
-
-/**
- * @brief Peripheral base class using Minres library
- *
- * @tparam regs_t Subclass of \ref peripheral_register_base that contains the
- *                register definitions.
- * @tparam owner_t Type of the owner pointer
- */
-template <typename regs_t, typename owner_t>
-class MemoryMappedPeripheralBase : public sc_core::sc_module,
-                             public PeripheralBase<owner_t>,
-                             public scc::tlm_target<> {
-  SC_HAS_PROCESS(MemoryMappedPeripheralBase);
-
- public:
-  class socket_accessor {
-  public:
-    constexpr socket_accessor(scc::tlm_target<>& parent) noexcept
-      : parent(parent) {}
-    socket_accessor(const socket_accessor&) = delete;
-    socket_accessor& operator=(const socket_accessor&) = delete;
-
-    tlm::tlm_target_socket<>* get() noexcept { return &parent.socket; }
-
-  private:
-    scc::tlm_target<>& parent;
-  };
-
-  using PeripheralBase<owner_t>::rst_in_;
-
-  MemoryMappedPeripheralBase(sc_core::sc_module_name name, PerParams&& params,
-                       owner_t* owner = nullptr)
-      : sc_core::sc_module{name}
-      , PeripheralBase<owner_t>{std::move(params), owner}
-      , scc::tlm_target<>{clk}
-      , NAMEDD(regs, regs_t) {
-    regs->registerResources(*this);
-    SC_METHOD(reset_cb);
-    sensitive << rst_in_;
-  }
-
-  /**
-   * @brief The socket to access the memory mapped registers of this peripheral
-   *
-   * Use sock_t_.get() to access a pointer to tlm::tlm_target_socket<>
-   */
-  socket_accessor sock_t_{*this};
-
- protected:
-  std::unique_ptr<regs_t> regs;
-  sc_core::sc_time clk;
-
-  void reset_cb() {
-    if (rst_in_.read())
-      regs->reset_start();
-    else
-      regs->reset_stop();
-  }
-};
-
-}  // namespace mmp
-
-#endif // __MMP_PERIPHERAL_BASE_H__
